RequestCollection holds request data.
The buffers are stored as `Vec<S>` and `Vec<R>`. In fact, the Vecs are
not formally pushed into or popped out of - the unmanaged memory at the
end of the buffer (capacity - len) is used.

These Vecs always have `len()` of `0`.

Remove scope from immediate_multiple_wait since ownership can be
statically verified now.